Add option for specifying columns to exclude from subset (#6)
After testing many different postgres data types and seeing that most of them are supported, but a small number are not, I've added the option to exclude certain columns from the subset. The idea would be that columns of custom data types which are not supported out-of-the box by the postgres jdbc driver (or whatever db jdbc driver is in use) would be handled manually by the user outside of the scope of the DBSubsetter project.

We may try to add better support for these data types in the future, but for now, the `--excludeColumns` option seems like the simplest workaround.