Test akka streams against single threaded version (#12)
Previously, only the single threaded version was under test. Now, both the single threaded version and the akka streams version are under test, and have the same test cases run against them.

The current setup is still quite messy and could still use a lot of work in terms of:
  * Saving time by only setting up origin DBs once instead of each test run
  * Adding extra protection by doing a complete checksum of each database, instead of just individual table checks
  * (Maybe?) Testing against the actual built JAR file for even more realistic tests
  * Including a performance test to catch major performance regressions

But, for now, this is at least safer than what we had before.