Workaround for MySQL Schema Introspection Bug (#23)
Workaround for the following bug:

MySQL schema introspection has a bug where they don't properly capitalize column names of the `pointedTo` side of foreign keys. It doesn't seem to be remedied by &useInformationSchema=true in DB URL so that the `DatabaseMetaDataUsingInfoSchema` class is used. This seems to be a bug at the MySQL layer, not at the JDBC Driver layer because the same issue is present in the command line program using the `show create table my_table` command

This is a hacky workaround and could cause problems in a schema where the same column name is used twice in the same table with different capitalization. (This seems like it ought to be either impossible or rare though). A less hacky workaround would be to match by column ordinal rather than by name, but unfortunately that info is not included in the result of the query we make to get foreign keys.

Now that there is a workaround for this bug, we are able to run the SchoolDB tests against MySQL, so this change also includes making the SchoolDB tests available for both MySQL and SQL Server.