Optimize memory by conditionally skipping Primary Key Storage (#28)
Add a manual configuration option to reduce memory consumption of DBSubsetter by skipping the PKStore for tables for which we are logically guaranteed to only see a given row once.

As of now, users need to understand which tables this applies to and configure this by hand. In a future release, we should implement a feature which will calculate and set the proper configuration automatically, rather than relying on user input.