Test many different primary key data types (#30)
Now that we are using Chronicle-Queue, we potentially have to serialize and write to disk any number of different types of primary keys (tinyint, uuid, unsigned integers, etc).

This adds automated tests for many of the most likely primary key types (including vendor-specific types). It also includes fixes for some of the problematic ones that the tests uncovered.