Favor composition over inheritance for ddl setup in Origin DB (#46)
Part 4 of a refactoring effort to make our testing infrastructure simpler and easier to work with in the future. In this part, we pause trying to extract things from AbstractEndToEndTest and focus instead on trying to reduce the large amount of inheritance which happens when individual tests inherit lots of Slick-related things they shouldn't need.

Specifically, instead of DDL being a trait so that each test class inherits all of its fields, DDL should instead be a class which can be composed as a single field in each test class that needs it.

This PR experiments with this approach by applying it to a single test -- the AutoIncrementPk tests.