Close sql connections after subsetting completes (#55)
Now that our e2e tests mostly all share the same docker containers, we are exhausting our connection limit to the databases. To fix this, this PR ensures that (most) of the SQL connections we open during a particular subsetting run will get closed once that run is done.

One SQL connection notably not managed by the `ConnectionFactory` we introduced in this PR lives in the `DbMetadataQueries` object. `close()` was already being called on it manually, so I did not bother addressing that here, although arguably eventually all SQL connections should be managed by the `ConnectionFactory`.