Test infra refactor faster load test origin setup (#56)
Makes strides to address https://github.com/bluerogue251/DBSubsetter/issues/39. While the code in this PR is still quite messy, and while because of that messiness it only currently supports PostgreSQL, still it's worth merging this now as it should make it significantly faster to run load tests, especially on subsequent runs.