Instrument pk store (#73)
Instrument PK Store to increment a Prometheus Counter whenever we discard a duplicate row from the origin DB (either when we tried to add it to the PK store or when we queried the PK store to check if it was already in it).

In the future we may also want to add a `Timer` onto the PK store methods to also get a sense of how fast or slow the PrimaryKeyStore is. But this PR does not include that. PrimaryKeyStore